+++ src/eth_node.py
@@ -7,6 +7,9 @@ class EthNode:
     def __init__(self, key: str): 
         self.w3 = Web3(Web3.HTTPProvider(f"https://mainnet.infura.io/v3/{key}"))    
 
+        self.makerT = "0x7dc5c0699ac8dd5250cbe368a2fc3b4a2daadb120ad07f6cccea29f83482686e"
+        self.takerT = "0x0fcf17fac114131b10f37b183c6a60f905911e52802caeeb3e6ea210398b81ab"
+
     # get timestamp of tx in unix time
     def getTimestamp(self, txHash: str | HexBytes) -> datetime:
         block = self.w3.eth.get_transaction(txHash)['blockNumber']
@@ -17,21 +20,20 @@ class EthNode:
     def getLogs(self, txHash: str | HexBytes) -> list:
         return self.w3.eth.get_transaction_receipt(txHash)['logs']
     
-    async def ws_v2_subscription_context_manager_example(self, q: asyncio.Queue, key: str):
+    async def w2Logs(self, q: asyncio.Queue, key: str):
         async with AsyncWeb3.persistent_websocket(
             WebsocketProviderV2(f"wss://mainnet.infura.io/ws/v3/{key}")
         ) as w3:
 
             await w3.eth.subscribe("logs", {
                 "address": "0xb2ecfE4E4D61f8790bbb9DE2D1259B9e2410CEA5",
-                "topics": ["0x7dc5c0699ac8dd5250cbe368a2fc3b4a2daadb120ad07f6cccea29f83482686e"]
-                }) # maker
+                "topics": [self.makerT]
+                }) 
             
             await w3.eth.subscribe("logs", {
                 "address": "0xb2ecfE4E4D61f8790bbb9DE2D1259B9e2410CEA5",
-                "topics": ["0x0fcf17fac114131b10f37b183c6a60f905911e52802caeeb3e6ea210398b81ab"]
-                }) # taker 
-
+                "topics": [self.takerT]
+                }) 
             async for response in w3.ws.process_subscriptions():
                 await q.put(response['result'])
